requirements.txt
Flask==3.0.3
gunicorn==22.0.0
numpy==1.26.4
Pillow==10.3.0
tensorflow==2.21.0
